refactor(bin): move `about` library to `/ui`
